small corrections to models and tests for models.WIP.
